+++ services/web/server/tests/unit/isolated/products/test_products_model.py
@@ -4,7 +4,6 @@
 # pylint: disable=too-many-arguments
 
 
-import json
 import re
 from itertools import chain
 from typing import Any
@@ -12,28 +11,16 @@ from typing import Any
 import pytest
 import simcore_service_webserver.products
 import sqlalchemy as sa
-from faker import Faker
 from models_library.basic_regex import TWILIO_ALPHANUMERIC_SENDER_ID_RE
 from models_library.products import ProductName
 from pydantic import BaseModel, ValidationError
-from pytest_simcore.helpers.faker_factories import random_product
 from pytest_simcore.pydantic_models import (
     assert_validation_model,
     walk_model_examples_in_package,
 )
 from simcore_postgres_database.models.products import products as products_table
-from simcore_service_webserver.constants import (
-    FRONTEND_APP_DEFAULT,
-    FRONTEND_APPS_AVAILABLE,
-)
+from simcore_service_webserver.constants import FRONTEND_APPS_AVAILABLE
 from simcore_service_webserver.products.models import Product
-from sqlalchemy import String
-from sqlalchemy.dialects import postgresql
-
-
-@pytest.fixture(scope="session")
-def product_name() -> ProductName:
-    return ProductName(FRONTEND_APP_DEFAULT)
 
 
 @pytest.mark.parametrize(
@@ -116,24 +103,6 @@ def test_product_host_regex_with_spaces():
     assert product.host_regex.search("osparc.bar.com")
 
     assert product.support_email == "foo@bar.com"
-
-
-@pytest.fixture
-def fake_product_from_db(faker: Faker, product_name: ProductName) -> dict[str, Any]:
-    server_defaults = {}
-    for c in products_table.columns:
-        if c.server_default is not None:
-            if isinstance(c.type, String):
-                server_defaults[c.name] = c.server_default.arg
-            elif isinstance(c.type, postgresql.JSONB):
-                m = re.match(r"^'(.+)'::jsonb$", c.server_default.arg.text)
-                if m:
-                    server_defaults[c.name] = json.loads(m.group(1))
-    return random_product(
-        name=product_name,
-        fake=faker,
-        **server_defaults,
-    )
 
 
 def test_safe_load_empty_blanks_on_string_cols_from_db(
